Fixed refresh problem where the subviews of lists, forms, trees and other views would not be resized (or relaid out) after the objects they are representing changed.
Fixed the refreshing of value fields so that multiple field views for the same value are all refreshed when one
of those fields is changed.

Added a second status bar to the bottom of the viewer that shows developer information.  Split up all status calls
into user based ones and developer based ones.  The developer's status bar can be turned off for normal use.

Added a description attribute to the UserAction class to take a description of an action.  These descriptions are
shown in the user's status bar (but are overriden if there is an About reason for vetoing the action).  Transfered over
the descriptions for most of the menu options.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006424 13f79535-47bb-0310-9956-ffa450edef68