Fixed two bugs:
Made the 'start of drag' event less sensitive, now you need to move four pixels vertically or horizontally before a drag starts

The initial drawing of the overlay view was occuring to above and to the right of the window that you were dragging, before snapping back to its correct location.  This was caused by the Drag event object having it cordinate changed during a call to pickup().  By reseting the position after this we have removed this problem.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006437 13f79535-47bb-0310-9956-ffa450edef68