Completely refactored the mouse event handling, redefining dragging within a view, content dragging, and moving views.  Events over borders are now changing the cursor properly, and mousing is more accurate.  Now the start of a drag action does not occur until the mouse has moved three pixels vertically or horizontally, which makes it easier to perform the click gestures without accidentally starting a drag.
Within tables, the resize is now done by a decorator on the table itself, and not by individual decorators on each of the cells.  The row heading is now also adjustable, and the column borders can be double-clicked on to resize to the needed width.

Empty fields now calculate their required size, so will show the complete title.

All views, except window borders, have had their simple borders replaced by ObjectBorder.  SimpleBorder gave each view a small handle on the right hand side on which the view menu could be invoked, and the view could be dragged.  The new border simply identifies the view and hightligh valid and invalid drop actions.  The view actions are still available but accessed by holding down the control key whilst right-clicking or dragging.

The workspace menu option has been pruned and reordered.

Double-clicking a class icon now brings up the fast finder dialog, rather than all instances.  All instances can accessed by holding the control key whilst double-clicking.  Both action can also be accessed from the icon's menu.

Removed the 'New Workspace' from all object menus except those that show an object that can be viewed properly within a workspace (specifically objects of type UserContext).

Class icons are now moved within the workspace, rather than creating a new instance.  This is consistent with moving aroud other views, and placing icons of objects on to the desktop.

When resizing the columns in a table or the size of a text field the limits of the possible size is shown during the drag rather than after the drag is complete.  Not the resize box will not get any smaller than an prespecified minimum size.

The left-hand pane of the browser frame is now resizable so that more of the right-hand pane can be seen.

Corrected the title for action dialogs brought up on classes so that the singular name is used rather than the fully qualified one.

All window views now have scrollbars so that when the window become larger than the workspace the rest of the view can be scrolled to.

Fixed bug where collection and object names are not displayed in the english forms, i.e. they had no spaces between words.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006494 13f79535-47bb-0310-9956-ffa450edef68