When a button is pressed on the button border any value field that is being edited is automatically before the button's action is invoked.
When dropping an object onto another object that does not have an 	action method taking that type, but has one or more fields of that type, the object will attempt to set up the first null field.

The window border's icon and title are now CONTENT area rather than VIEW area, and will allow the object menu to be accessed or for the object reference to be dragged elsewher.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006496 13f79535-47bb-0310-9956-ffa450edef68