Drop-downs (lookup) on already set fields now replace the current selection with the new selection.
Programmatically updated references within objects are now updated on screen.

Value parameter fields in action dialogs are now accepting default values, and allowing the user to change them.  Previously the fields would default to their NakedValue type's default just before the method was executed.

Fixed bug in debug view for the overlay view, which prevented the debug from displaying.

Removed the drop-down arrow from the drop-down selection list.

Adjusted the position of the drop-down arrow so it wasn't on the border of the empty/icon field

Fixed the painting so that only the damaged area of the UI is repainted, saving from the expenses of repainting the entire screen everytime the mouse moved.


git-svn-id: https://svn.apache.org/repos/asf/incubator/isis/trunk@1006537 13f79535-47bb-0310-9956-ffa450edef68