ISIS-309, ISIS-313: updates to get gmap3 to work; example fix
Rather a lot of faff, principally to do with getting the necessary Javascript and CSS contributions
to be contributed to the header even though the object is not initially rendered.

In the end, I ended up rendering them all, but using CSS to hide the components that
aren't meant to be there (in LinksSelectorPanelAbstract).

To avoid accidentally resolving a lazily loaded colleciton, also introduced
the notion of a "dummy model"... an equivalent model that is empty, and which is passed into
the hidden components.  The subclasses of LinksSelectorPanelAbstract are required
to provide this dummy model.  In the case of Collections, the
CollecitonContentsLinksSelectorPanel delegates off to its model, EntityCollectionModel,
to create the dummy model.

The JDO example has also been updated to have the gmap3 code, but is
commented out.  Intention is that a tutorial can show how to wire in
new components just by commenting back in the relevant chunks of stuff.

~~~
This commit also fixes a bug in the ToDoItem (ISIS-313)