ISIS-554: dashboard service (@Dashboard annotation)
A single domain service can be annotated with @Dashboard; the Wicket
viewer will then render in the same way as a domain object.  Note that for
the dashboard service the usual restrictions on having no properties,
collections and contributed actions have been relaxed.

The ToDo app has also been refactored to demonstrate the use of the dashboard,
with an additional new analysis view model.

Additionally for the ToDo app:
* icons for ToDoItem improved, also distinguishing done vs todo items
* a ClockService has been added

Known issue:
* the Restful Objects viewer cannot render the properties/collections of services.