ISIS-554: reworking dashboard into a @HomePage annotation
instead of a @Dashboard service, instead we indicate that precisely (at most)
one domain service action is the @HomePage action.  This can return any object,
eg a view model representing a dashboard, or a simple list, or may even
accept arguments (in which case an action prompt will be automatically displayed).

In addition:
- slight hack to lazily maintain the specId cache after loading any spec
  - required for view model specs that are not eagerly loaded
- new convenience AbstractHomePageDashboardService added to applib
- ActionPanel's NoResultHandler removed, since had no function
  (and was causing the EmptyCollectionPanelFactory to not correctly render an empty
  collection for @HomePage action that returned no results)
- added convenience AbstractContainedObject#newViewModelInstance(...) method

Minor fixes to CSS:
- so that hovering over a plain label in a Wicket viewer
  table is shown as a regular default cursor, not a pointer (hand)
- on custom application.css in quickstart app, for x-abc demo styles

- minor fixes to CSS