ISIS-401: show concurrency ex. if object changed while action prompt is displayed
In fact, was displaying already for some actions (eg ToDoItem#updateCost).  However,
for ToDoItem#updateCategories(...) it was not, because the loading of the
object adapters to compute the choices was causing the adapter to be
silently resynced.

This is now captured and we forward onto the entity page, as per normal.