ISIS-719: use webjars for 3rd party css and jquery.
For JQuery we use Wicket's version
For bootstrap, we seem to be using a slight custom version compared to the vanilla version.

So, although this commit adds in the infrastructure for webjars, it doesn't actually take use of it.

Even so, have still decided to commit, since may be of use in future for other dependencies (or if we decide to use a different version of jQuery than the one that is bundled with Wicket).