ISIS-776: CompositeFixtureScript hierarchy evaluated lazily.
Previously CompositeFixtureScript#addChildren was called prior to the execution, meaning that the hierarchy of scripts (at least so far as any given CompositeFixtureScript was concerned) was specified before execution.

This has now been refactored so that the hierarchy is defined dynamically within the execute() method itself...
- the old doRun(String parameters) is renamed execute(parameters)
- the old add(FixtureScript fs) is replaced by execute(FixtureScript fs, parameters).

This also allows parameters to be specified dynamically if required.  All together a more flexible design.