ISIS-781: editable view models via ViewModel.Cloneable interface.
For those view models that implement Cloneable the wicket viewer is able to simulate editable objects by replacing the adapter memento held within the EntityModel with a new one created by cloning the mutated state of the initial view model.

Demonstrated using a new ToDoItemWizard in the todo (quickstart) app.