ISIS-889: tweaked the javascript to adjust margin.
The action prompt dialog is *meant to* be sized through our call to Wicket.Window.get().autoSizeWindow().  However, it seems not to size correctly on Chrome.  I noticed though that manually adjusting the margin on one of the aciton panel's container element down by a pixel sorted out the issue.

So this is pretty hacky, but I've added in a patch to the javascript that calls ww.autoSizeWindow(), to manually change the CSS by 1 pixel.  Seems to do the job, and unnoticeable on Firefox.