ISIS-948: refactoring the EventBusService (and EventBusServiceDefault subclass) so that the guava EventBus is actually instantiated only once, in effect with singleton scope, rather than with request-scoped.
Although this design is not symmetrical (in the various @PostConstruct and @PreDestroy lifecycle methods), it has many fewer moving parts.  Making the guava EventBus a singleton means that all events flow through this single instance.  We rely on the proxies of the request-scoped services to dispatch to the appropriate instance on the thread (assuming that a transaction is running for that thread).