ISIS-1000: generalizing CssClassFacet to accept an (optional) ObjectAdapter
This allows the same facet to be used for the object-level cssClass() (corresponding to CssClassFacetMethod) as well as the member-level @XxxLayout(cssClass=...) annotations.  It also opens up the possibility of member-level CSS being dynamic (per object instance) in the future.

Also, updated todoapp to demonstrate the feature (completed items are rendered in tables with a line-through and lightgrey).