ISIS-1189: choices and autoComplete facets (for both properties and action parameters) now automatically filter out any objects that are not visible to the current user.
to do this they now accept an AuthenticationSession and DeploymentCategory so that they can instantiate a VisibilityContext and call InteractionUtils against the appropriate facetholder (usually that of the type of the property or parameter).

The (Wicket and RO) viewers are responsible for providing the AuthenticationSession and DeploymentCategory.  In the case of the Wicket viewer, this is looked up by the UI components (eg ReferencePanel) from the IsisContext.  In the case of the RO viewer, this is passed through by the ResourceContext / RendererContext (a new subinterface to ensure backward compatibility with any custom extensions to the RepresentationService).