ISIS-993: ScopedSessionAttribute now using Component rather than HasPath.  That means that the state of the page (selected tabs, selected collection view) is being persisted and honoured in the session correctly.
Also added in new implementations of UiHintPathSignificant, however this bit is incomplete... hacked the fullHintPathFor(...) to ignore the fact as to whether a component is "UI Hint path significant" or not.  Intention in next commit is to simplify and get rid of this interface, just use all the components in the page as being significant.