ISIS-1427: renames methods in IsisSystemForTest, rationalizes functionality between IsisSystem and IsisSessionFactory
The intent being that IsisSystem is just a builder for IsisSessionFactory.  The responsibility for calling @PostConstruct and @PreDestroy was held by IsisSystem, meaning that it was necessary to call both shutdown on IsisSessionFactory and then on IsisSystem.  Moving the responsibility for services initialization/destroy into IsisSessionFactory therefore simplifies the callers; there is no need to hold onto the IsisSystem, just use the IsisSessionFactory.

As part of this the DeploymentCategoryProviderDefault is now a fallback service, added into ServicesInjector prior to init'ing the guice IsisInjectModel (which creates the IsisSystem and thence IsisSessionFactory).