ISIS-1448: reinstates the ability to switch component factory for entity collection.
Prior to 1.12.0 the ComponentFactory SPI was used to lookup the component to render an entity collection, using ComponentType.ENTITY_COLLECTION and passing in an EntityModel with appropriate hints to indicate which collection of the entity to render.  The default implementation was EntityCollectionsPanelFactory returning an EntityCollectionsPanel (the plural here is because there many representations of a single collection, with a drop-down selector to choose which).

In the 1.12.0 (layout XML stuff), this capability was lost: the new Col class simply instantiates EntityCollectionPanel (renamed from EntityCollectionsPanel).

This commit reinstates the ability by having Col use the ComponentFactoryRegistrar to look up the component to render the entity collection.

The commit also deletes the CollectionPanelFactory; this code was unused.  CollectionPanel is the default table representation of the collection; EntityCollectionPanel simply instantiates it and then keeps track of all the "other" implementations to render the collection.  To make this easier to switch in, the code to instantiate a CollectionPanel within EntityCollectionPanel has been moved to a (potentially overridable) factory method.