ISIS-1454: fixes collection hints in the Wicket viewer.
Various elements of each page - tab, page number, showAll and collection selector - can be set by hints taken from either the URL or from the session (via HintStore).  The hint key is based on the (Wicket) component path plus additional qualifier, eg 'showAll'.

This commit fixes an issue with the collection selector and the page number hints, which were not being honoured.  The problem was down to passing in the incorrect component in some cases... now always passes in the CollectionSelectorPanel.