ISIS-1571: checks that every action collection parameter has a supporting choices or autoComplete facet.
Implementing this required moving the metamodel validation to be run within a session.  This results in potentially more object specs to validate, and in particular resulted in a validation failure that java.lang.Object and FreeStandingList had the same object type.  So this commit also incorporates a fix for that, by only comparing specs that are entities or view models.