ISIS-1613: fixes NPE with bulk actions
which was resulting from the actionModel hint, which holds the action to invoke to recreate the list of domain objects, was caching its results and so was returning a list of stale pojos that were implicitly detached and therefore mapped as transient.

Fix was simply to invalidate this cached return result by invalidating the actionModel.