ISIS-1630: no longer installs DN/JDO facets on view models (or rather, any object that isn't an entity).
The issue as originally described was the result of the MandatoryFromJdoColumnAnnotationFacetFactory being installed, with the MandatoryFacet with inverted semantics (due to absence of @Column) trashing the original MandatoryFacetDefault that was already there.  The fix is to simply ignore the facet factory for

The other facet factories pertaining only to DN entities have also been updated in same way.