ISIS-1750: adds fallback to previous behaviour if fail to handle exception gracefully.  This can occur if attempt to access non-existent object....
This can occur if attempt to access non-existent object.  An exception is thrown which takes us to the OnException handler, and then the respondGracefully for the recognised exception again attempts to take us back to the entity page which promptly throws the same exception once more.

So, if this happens we just give up and revert to the original behaviour of showing the error page.