ISIS-1569: continues if replicated an exception; utility methods for DTOs
ContentMappingServiceForCommandDto now takes responsibility for automatically copying over 5 fields from the CommandWithDto (entity) into the user data of the CommandDto.
One of these is the exception, if any.  This is used by BackgroundCommandExecution to CONTINUE to process EVEN IF an exception occurs on the slave, if an exception also had occurred on the master.