ISIS-898: fixes classcast exception with TreePanel.
The FixtureResult class (view model returned from running fixture scripts through the UI) has a property of type java.lang.Object - this being the object wrapped by the FixtureResult, created by the fixture script.  The logic in TreePanelFactories to determine if this property's type implements TreeNode seems to be wrong - returns yes for Object being a subtype of TreeNode.  This then causes a TreePanel to attempt to be rendered which looks to downcast the object to TreeNode, and then that fails.