ISIS-2020: fixes the listAll issue
I believe, that the issue is on DN's side with the PersistenceManager
not creating queries that act upon the correct DB connection.

As of DN-5.2 FederatedJDOQLQuery is still work in progress.

We fixed this by utilizing the new type-safe queries as provided with
DN-5, which appear do work correctly on Federated Data Stores.

As a side-effect we fixed the 'disable-MultivaluedFetch' optimization,
which was not working in our implementation since ther was a key
renaming in DN's migration from 4.1.0.M1 to 4.1.0.M2


Task-Url: https://issues.apache.org/jira/browse/ISIS-2020