ISIS-2218: AuthorizationManagerDefault renamed to AuthorizationManager, removes interface
In the process, moved from RuntimeServices to SecurityApi.

To do this required (re)moving te MetaModelRefiner implementation from AuthorizationManagerDefault, because that has a dependency on metamodel (and security-api does not depend on metamodel).  However, since there is now only ever one implementation of AuthorizationManager, the adding of the AuthorizationFacetFactory is now simply part o fthe ProgrammingModelJava8.

Also:
- inlined MetaModelRefiner.getAll ... only used in one place.
- removes init() and shutdown() from Authorizor API, none of the implementations used them (and they can always just declare a @PostConstruct if nec).
- renames XxxAuthorizor to AuthorizerXxx and register as @Service's rather than @Beans

And:
- renames IsisModuleSecurityRealm to IsisModuleExtSecmanShiroRealm