ISIS-1922: Notify user when expired session is refreshed
Without this patch applied, when a user clicks a menu item in an expired
session, the session is simply refreshed and no action is taken. This is
a problem because from the user's point of view the click is ignored. This
patch fixes the problem by notifying user that the session was expired and
was refreshed from the click.