ISIS-2758: reworks ImpersonatedUserHolderUsingHttpSession, so not @RequestScoped
as this prevents it from being used in initial fixture scripts.

domain-app-demo/persist-all/simple-object-builder: Error creating bean with name 'scopedTarget.isis.webapp.ImpersonatedUserHolderUsingHttpSession': Scope 'request' is not active for the current thread; consider defining a scoped proxy for this bean if you intend to refer to it from a singleton; nested exception is java.lang.IllegalStateException: No thread-bound request found: Are you referring to request attributes outside of an actual web request, or processing a request outside of the originally receiving thread? If you are actually operating within a web request and still receive this message, your code is probably running outside of DispatcherServlet: In this case, use RequestContextListener or RequestContextFilter to expose the current request.