CAUSEWAY-3553: adds support to fix JDK xml pretty printing bug
in JDK 11 (https://bugs.openjdk.org/browse/JDK-8262285), taking inspiration from the fix in the Approvals library.

In essence, if the JaxbOptions are set to produce indented output (the default), the we additional use XSLT
to pretty print it.  The TransformerFactory used there can additionally be customised, but it also defaults
to using an indent of 4